+++ KotOR_IO/File Formats/GFF FieldTypes/6_DWORD64.cs
@@ -1,0 +1,72 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.IO;
+
+namespace KotOR_IO
+{
+    public partial class GFF
+    {
+        public class DWORD64 : FIELD
+        {
+            public ulong value;
+
+            public DWORD64() { }
+            public DWORD64(BinaryReader br, int offset)
+            {
+                //Header Info
+                br.BaseStream.Seek(24, 0);
+                int LabelOffset = br.ReadInt32();
+                int LabelCount = br.ReadInt32();
+                int FieldDataOffset = br.ReadInt32();
+
+                //Basic Field Data
+                br.BaseStream.Seek(offset, 0);
+                Type = br.ReadInt32();
+                int LabelIndex = br.ReadInt32();
+                int DataOrDataOffset = br.ReadInt32();
+
+                //Label Logic
+                br.BaseStream.Seek(LabelOffset + LabelIndex * 16, 0);
+                Label = new string(br.ReadChars(16)).TrimEnd('\0');
+
+                //Comlex Value Logic
+                br.BaseStream.Seek(FieldDataOffset + DataOrDataOffset, 0);
+                value = br.ReadUInt64();
+
+
+            }
+
+            internal override void collect_fields(ref List<Tuple<FIELD, int, int>> Field_Array, ref List<byte> Raw_Field_Data_Block, ref List<string> Label_Array, ref int Struct_Indexer, ref int List_Indices_Counter)
+            {
+                Tuple<FIELD, int, int> T = new Tuple<FIELD, int, int>(this, Raw_Field_Data_Block.Count, this.GetHashCode());
+                Raw_Field_Data_Block.AddRange(BitConverter.GetBytes(value));
+                Field_Array.Add(T);
+
+                if (!Label_Array.Contains(Label))
+                {
+                    Label_Array.Add(Label);
+                }
+            }
+
+            public override bool Equals(object obj)
+            {
+                if ((obj == null) || !GetType().Equals(obj.GetType()))
+                {
+                    return false;
+                }
+                else
+                {
+                    return value == (obj as DWORD64).value && Label == (obj as DWORD64).Label;
+                }
+            }
+
+            public override int GetHashCode()
+            {
+                return new { Type, value, Label }.GetHashCode();
+            }
+
+
+        }
+    }
+}
